Validate the number of instances and provide a wrapper for nodetool.

+++ cassandra/cassandra.yaml.tpl
@@ -197,18 +197,25 @@ write_files:
     }
 
     if [ ! -f "/etc/cassandra/.configured" ]; then
-      echo "The Cassandra instances are already configured. To reconfigure, please remove the /etc/cassandra/.configured file."
-      exit
-    fi
-
-    if [[ "$seed_host" == "127.0.0.1" ]]; then
-      echo "Please specify with seed_host."
-      exit
+      echo "Error: the Cassandra instances have not been configured. Please run /etc/cassandra/configure_cassandra.sh"
+      exit 1
     fi
 
     if [ "$(id -u -n)" != "root" ]; then
       echo "Error: you must run this script as root" >&2
       exit 4  # According to LSB: 4 - user had insufficient privileges
+    fi
+
+    re='^[0-9]+$'
+    available=$(ip a | grep "^[0-9]: eth" | wc -l)
+    if ! [[ $instances =~ $re ]] || [[ $instances > $available ]] || [[ $instances < 1 ]]; then
+      echo "Error: please provide a number of instances between 1 and $available"
+      exit 1
+    fi
+
+    if [[ "$seed_host" == "127.0.0.1" ]]; then
+      echo "Error: please specify with seed_host."
+      exit 1
     fi
 
     # Ensure the init.d version is not usable by an operator
@@ -295,8 +302,8 @@ write_files:
       repo_ver="40x"
     fi
     if [[ "$repo_ver" == "" ]]; then
-      echo "Error: Only Cassandra 3.11.x or 4.0.x are supported".
-      exit
+      echo "Error: only Cassandra 3.11.x or 4.0.x are supported".
+      exit 1
     fi
 
     # The archive repository contains all versions of Cassandra
@@ -351,6 +358,13 @@ write_files:
       exit 4  # According to LSB: 4 - user had insufficient privileges
     fi
 
+    re='^[0-9]+$'
+    available=$(ip a | grep "^[0-9]: eth" | wc -l)
+    if ! [[ $instances =~ $re ]] || [[ $instances > $available ]] || [[ $instances < 1 ]]; then
+      echo "Error: please provide a number of instances between 1 and $available"
+      exit 1
+    fi
+
     directories=("commitlog" "data" "hints" "saved_caches")
     data_location=/var/lib/cassandra
     log_location=/var/log/cassandra
@@ -361,7 +375,7 @@ write_files:
 
     echo "Waiting for $instances disks to be available"
     while [ $(ls -l /dev/disk/azure/scsi1 | grep lun | wc -l) -lt $instances ]; do
-        printf '.'
+      printf '.'
       sleep 10
     done
 
@@ -436,6 +450,13 @@ write_files:
       exit 4  # According to LSB: 4 - user had insufficient privileges
     fi
 
+    re='^[0-9]+$'
+    available=$(ip a | grep "^[0-9]: eth" | wc -l)
+    if ! [[ $instances =~ $re ]] || [[ $instances > $available ]] || [[ $instances < 1 ]]; then
+      echo "Error: please provide a number of instances between 1 and $available"
+      exit 1
+    fi
+
     echo "Configuring rsyslog..."
     rsyslog_file=/etc/rsyslog.d/cassandra.conf
     rm -f $rsyslog_file
@@ -480,7 +501,7 @@ write_files:
     }
 
     if [ -f "/etc/cassandra/.configured" ]; then
-      echo "Warning: Cassandra instances already configured."
+      echo "Warning: cassandra instances already configured."
       exit
     fi
 
@@ -489,9 +510,16 @@ write_files:
       exit 4  # According to LSB: 4 - user had insufficient privileges
     fi
 
+    re='^[0-9]+$'
+    available=$(ip a | grep "^[0-9]: eth" | wc -l)
+    if ! [[ $instances =~ $re ]] || [[ $instances > $available ]] || [[ $instances < 1 ]]; then
+      echo "Error: please provide a number of instances between 1 and $available"
+      exit 1
+    fi
+
     if [[ "$seed_host" == "127.0.0.1" ]]; then
-      echo "Error: Please specify with seed_host."
-      exit
+      echo "Error: please specify with seed_host."
+      exit 1
     fi
 
     version=$(rpm -q --queryformat '%%{VERSION}' cassandra)
@@ -610,11 +638,46 @@ write_files:
       shift
     done
 
+    re='^[0-9]+$'
+    available=$(ip a | grep "^[0-9]: eth" | wc -l)
+    if ! [[ $instances =~ $re ]] || [[ $instances > $available ]] || [[ $instances < 1 ]]; then
+      echo "Error: please provide a number of instances between 1 and $available"
+      exit 1
+    fi
+
     for i in $(seq 1 $instances); do
       intf="eth$(expr $i - 1)"
       ipaddr=$(ifconfig $intf | grep 'inet[^6]' | awk '{print $2}')
       nodetool -u cassandra -pw cassandra -h $ipaddr -p 7$${i}99 setstreamthroughput -- $throughput
     done
+
+- owner: root:root
+  permissions: '0755'
+  path: /etc/cassandra/nodetool.sh
+  content: |
+    #!/bin/bash
+
+    instance=""
+    if [[ $1 == *"--"* ]]; then
+      param="$${1/--/}"
+      declare $param="$2"
+    fi
+    shift
+    shift
+
+    re='^[0-9]+$'
+    available=$(ip a | grep "^[0-9]: eth" | wc -l)
+    if ! [[ $instance =~ $re ]] || [[ $instance > $available ]] || [[ $instance < 1 ]]; then
+      echo "Error: please provide an instance number between 1 and $available"
+      exit 1
+    fi
+
+    intf="eth$(expr $instance - 1)"
+    ipaddr=$(ifconfig $intf | grep 'inet[^6]' | awk '{print $2}')
+
+    echo "Instance $instance"
+    echo "Issuing nodetool -h $ipaddr -p 7$${instance}99 $@"
+    nodetool -u cassandra -pw cassandra -h $ipaddr -p 7$${instance}99 $@
 
 - owner: root:root
   permissions: '0400'
